refactor(registration): extract remaining admin operations
- Extract handleClearAllCourts -> handleClearAllCourtsOp
- Extract handleBlockCreate -> handleBlockCreateOp
- Extract handleCancelBlock -> handleCancelBlockOp
- Extract handleAdminClearCourt -> handleAdminClearCourtOp
- Extract handleMoveCourt -> handleMoveCourtOp
- Extract handleReorderWaitlist -> handleReorderWaitlistOp
- All thin wrappers preserve original function names
